+++ docs/examples/customization/filters.js
@@ -1,6 +1,6 @@
 angular.module('selectDemo')
     .filter('mySearchFilter', ['$sce', function($sce) {
-        return function(label) {
+        return function(label, query, option, element) {
 
             var html = '<i>' + label + '</i>';
 
@@ -9,7 +9,7 @@ angular.module('selectDemo')
     }])
 
     .filter('myDropdownFilter', ['$sce', function($sce) {
-        return function(label, query, option) {
+        return function(label, query, option, element) {
 
             var html = '<kbd>#' + option.id + '</kbd> ' + label;
 
@@ -18,13 +18,30 @@ angular.module('selectDemo')
     }])
 
     .filter('myListFilter', function() {
-        return function (list, query, getLabel) {
+        return function (list, query, getLabel, options, element) {
             return list;
         }
     })
 
     .factory('myEditItem', function() {
-        return function(removedValue, lastQuery, getLabel) {
+        return function(removedValue, lastQuery, getLabel, editItemIsCorrected, element) {
             return removedValue ? getLabel(removedValue) + ' :-)' : '';
+        };
+    })
+
+    .filter('myGroupFilter', function($sce) {
+        return function(label, query, items, options, element) {
+
+            element[query ? 'addClass' : 'removeClass']('show');
+
+            function toggle(element){
+                angular.element(element).parent().parent().find('li').toggleClass('show');
+                var iconElement = angular.element(element).find('span');
+                iconElement.html(iconElement.html() === '&plus;' ? '&minus;' : '&plus;');
+            }
+
+            var html = '<div class="group-header" onclick="(' + toggle.toString() + ')(this)"><span class="group-header-plus">&plus;</span> ' + label + '</div>';
+
+            return $sce.trustAsHtml(html);
         };
     });